Rename core/ to frameworks/, extract BaseEccentricHM base class
