+++ gwModels/__init__.py
@@ -1,4 +1,4 @@
-from . import core
+from . import frameworks
 from . import models
 from . import utils
 from . import ecc_measures
