--- gwModels/core/__init__.py
@@ -1,2 +1,0 @@
-from .gwnrhme import *
-from .gwnrxhme import *
