+++ gwModels/frameworks/__init__.py
@@ -1,0 +1,3 @@
+from .base import BaseEccentricHM
+from .gwnrhme import NRHME
+from .gwnrxhme import NRXHME
